requirements.txt
python-telegram-bot[job-queue]>=20.0
yt-dlp>=2024.1.0
openai-whisper>=20231117
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
aiofiles>=23.0.0
tenacity>=8.2.0
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
httpx>=0.26.0
Pillow>=10.0.0
pytesseract>=0.3.10
supabase>=2.0.0
